First commit of a Bag class.
This is not a complete implementation, but it is okay
to commit partial work to a git repository.  Whether one should commit partial work depends on project or corprorate policy.  If a repository has a production branch, one would not push incomplete code to the production branch. For purposes of this class, the main branch of csci-356 is a development branch, and thus we can commit partial code to it.