First commit of the MSDie example.
msdie.py is based on code from Chapter 2 in "Problem Solving with Algorithms
and Data Structures using Python."

  https://runestone.academy/ns/books/published/pythonds/ProperClasses

I extended this example with type hints, method comments with restructuredText,
parameter and return type documentation, adding 'if __name__ == "__main__"',
and moving the code at the end of msdie.py that provides example usage of MSDie
into a function named main().

The usage of 'if __name__ == "__main__"' and moving examples into a main()
function allow MSDie to be imported from msdie.py into other modules without
causing the code at the end of msdie.py to be executed.