Code used in class in Lecture 10.
amortized_append.py looks at the execution time of appends to Python's
built-in list.

simple_list.py plots capacity ofa "simple" list implementation
as a function of the number of appends.  The implementation is simple
in that it doesn't support all of the operations of the Python
built-in list and it is implemented completely in Python rather than
C.  I do this so that we have an example of code for performing
an amortized time complexity analysis of our SimpleList's append function.

list_caps.py plots the capacity (memory allocated) to a Python
built-in list containing n items..

my_dict.py plots the execution time for key-value lookups into a dict.